- apply CRAI suggestions + refactor of ServiceFacade into BaseService to avoid misuse of Facade by other services types not needing lifecycle aware methods
