+++ androidNode/src/androidMain/kotlin/network/bisq/mobile/android/node/service/network/NodeConnectivityService.kt
@@ -14,14 +14,6 @@ class NodeConnectivityService(private val applicationService: AndroidApplication
         const val SLOW_PEER_QUANTITY_THRESHOLD = 4
     }
 
-    override fun activate() {
-        super.activate()
-    }
-
-    override fun deactivate() {
-        super.deactivate()
-    }
-
     override fun isConnected(): Boolean {
         val connections = currentConnections()
         log.v { "Connected peers = $connections" }
